+++ rust+yew/client/src/models/entity_subset.rs
@@ -3,5 +3,5 @@ use serde::Deserialize;
 #[derive(Deserialize)]
 pub struct EntitySubset<T> {
     pub items: Vec<T>,
-    pub total: i64,
+    pub total: i32,
 }
