+++ rust+yew/server/src/models/entity_subset.rs
@@ -3,5 +3,5 @@ use serde::Serialize;
 #[derive(Serialize)]
 pub struct EntitySubset<T> {
     pub items: Vec<T>,
-    pub total: i64,
+    pub total: i32,
 }
